+++ platform/3ds/source/hardware_profile.h
@@ -12,7 +12,7 @@ typedef struct Platform3DSHardwareProfile {
 
 static inline const Platform3DSHardwareProfile *Platform3DS_ProfileForModel(bool is_new) {
   static const Platform3DSHardwareProfile old_profile = {
-    "Old 3DS E8", true, true, true, true
+    "Old 3DS E9", true, true, true, true
   };
   static const Platform3DSHardwareProfile new_profile = {
     "New 3DS E6 renderer", false, false, false, false